move `peek_n` to utils
