+++ relay-server/src/utils/stream/mod.rs
@@ -1,7 +1,9 @@
 mod bounded;
 mod metered;
+mod peek;
 mod retryable;
 
 pub use bounded::*;
 pub use metered::*;
+pub use peek::*;
 pub use retryable::*;
